requirements.txt
Django==2.1.5
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
Markdown==3.1.1
Pillow==5.2.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2018.9
